refactor(connectors): consolidate m365 shared code into _m365/
Promote duplicated MSAL auth, markdown→HTML, and Graph error mapping
from teams/ and outlook/ into src/connectors/_m365/ — used by both
microsoft connectors via M365Auth / loadM365Credentials / M365Error /
makeM365MapError. Public surface preserved via aliased re-exports
(TeamsAuth/OutlookAuth, loadTeamsCredentials/loadOutlookCredentials, etc.).

CODE_MAP is the union of teams' and outlook's: outlook's four extra
Exchange-flavored codes (ErrorAccessDenied, ErrorItemNotFound,
ApplicationThrottled, ErrorInvalidIdMalformed) are now visible to both
connectors. No behavior change — they map to canonical codes either
way; extra keys only get hit when Graph returns them.

- net -648 lines (-677 / +29)
- 6 duplicated files deleted across teams/lib and outlook/lib
- 3 new files in _m365/ (auth.ts, error.ts, markdown_to_html.ts)
- npx tsc --noEmit clean
